# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_module_collaboration.CreateTeamRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      visibility:
        allOf:
        - $ref: '#/definitions/internal_module_collaboration.Visibility'
        enum:
        - public
        - private
    required:
    - name
    type: object
...
  internal_module_collaboration.TeamResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      member_count:
        type: integer
      member_limit:
        type: integer
      my_role:
        allOf:
        - $ref: '#/definitions/internal_module_collaboration.Role'
        description: 'Optional: current user''s role in this team'
      name:
        type: string
      owner_id:
        type: string
      slug:
        type: string
      updated_at:
        type: string
      visibility:
        $ref: '#/definitions/internal_module_collaboration.Visibility'
    type: object
...
  internal_module_collaboration.Visibility:
    enum:
    - public
    - private
    type: string
    x-enum-varnames:
    - VisibilityPublic
    - VisibilityPrivate
...
host: localhost:8080
info:
  contact:
    email: support@uniedit.io
    name: UniEdit Support
    url: https://uniedit.io/support
  description: UniEdit 视频编辑器后端 API
  license:
    name: Proprietary
    url: https://uniedit.io/license
  termsOfService: https://uniedit.io/terms
  title: UniEdit Server API
  version: "1.0"
paths:
...
  /invitations:
    get:
      description: Get all pending invitations for the current user
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List my invitations
      tags:
      - Collaboration
  /invitations/{id}:
    delete:
      description: Revoke a pending invitation
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Revoke invitation
      tags:
      - Collaboration
  /invitations/{token}/accept:
    post:
      description: Accept a team invitation
      parameters:
      - description: Invitation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "410":
          description: Gone
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Accept invitation
      tags:
      - Collaboration
  /invitations/{token}/reject:
    post:
      description: Reject a team invitation
      parameters:
      - description: Invitation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject invitation
      tags:
      - Collaboration
...
  /teams:
    get:
      description: Get all teams the current user belongs to
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List my teams
      tags:
      - Collaboration
    post:
      consumes:
      - application/json
      description: Create a new team
      parameters:
      - description: Create team request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_module_collaboration.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_module_collaboration.TeamResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create team
      tags:
      - Collaboration
  /teams/{slug}:
    delete:
      description: Permanently delete a team
      parameters:
      - description: Team slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete team
      tags:
      - Collaboration
    get:
      description: Get team details by slug
      parameters:
      - description: Team slug
        in: path
        name: slug
        required: true
        type: string
      - description: Owner ID
        in: query
        name: owner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_collaboration.TeamResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get team
      tags:
      - Collaboration
    patch:
      consumes:
      - application/json
      description: Update team settings
      parameters:
      - description: Team slug
        in: path
        name: slug
        required: true
        type: string
      - description: Update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_module_collaboration.UpdateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_collaboration.TeamResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update team
      tags:
      - Collaboration
...
  /teams/{slug}/leave:
    post:
      description: Leave a team (current user)
      parameters:
      - description: Team slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Leave team
      tags:
      - Collaboration
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"